Added native tooltips to charts, improved mobile styling
