+++ resources/pages/Profile/components/AddFlight/index.ts
@@ -1,0 +1,1 @@
+export * from './AddFlightForm';
