+++ resources/pages/Profile/components/CurrentFlight/constants.ts
@@ -1,0 +1,25 @@
+import type { FlightDelayStatus } from '../../types';
+
+export const TEXT_COLORS: Record<FlightDelayStatus, string> = {
+  none: 'text-success',
+  moderate: 'text-warning',
+  severe: 'text-error',
+};
+
+export const CARD_COLORS: Record<FlightDelayStatus, string> = {
+  none: 'bg-success/10',
+  moderate: 'bg-warning/10',
+  severe: 'bg-error/10',
+};
+
+export const CARD_BORDER_COLORS: Record<FlightDelayStatus, string> = {
+  none: 'border-success/20',
+  moderate: 'border-warning/20',
+  severe: 'border-error/20',
+};
+
+export const PROGRESS_BAR_COLORS: Record<FlightDelayStatus, string> = {
+  none: 'progress-success',
+  moderate: 'progress-warning',
+  severe: 'progress-error',
+};
