+++ resources/pages/Profile/components/CurrentFlight/index.ts
@@ -1,0 +1,1 @@
+export * from './CurrentFlightCard';
