+++ resources/pages/Profile/components/Map/CesiumMap.tsx
@@ -2,9 +2,9 @@ import { Cartesian3, Color, type Viewer as ViewerType } from 'cesium';
 import { useRef, type Dispatch, type SetStateAction, useEffect } from 'react';
 import { type UseFormReturn, useWatch } from 'react-hook-form';
 import { type CesiumComponentRef, Entity, Viewer } from 'resium';
-import planeIconUrl from '../../assets/plane.svg';
+import planeIconUrl from '../../../../assets/plane.svg';
 import { type MapCardFormData } from './MapCard';
-import { type MapFlight, type FilteredMapData, type MapCoords } from './utils';
+import type { FilteredMapData, MapCoords, MapFlight } from './types';
 
 export interface CesiumMapProps {
   center: MapCoords;
