+++ resources/pages/Profile/components/Map/constants.ts
@@ -1,0 +1,4 @@
+export const DEFAULT_COORDINATES = {
+  lat: 0,
+  lng: 0,
+};
