+++ resources/pages/Profile/components/RecentFlights/index.ts
@@ -1,0 +1,2 @@
+export * from './CompletedFlights';
+export * from './UpcomingFlights';
