+++ resources/pages/Profile/components/Statistics/TopAirportsChart.tsx
@@ -2,26 +2,23 @@ import { ResponsiveBar } from '@nivo/bar';
 import classNames from 'classnames';
 import { useForm, useWatch } from 'react-hook-form';
 import { useParams } from 'react-router-dom';
-import { Form, Loading, Select } from 'stratosphere-ui';
-import { trpc } from '../../utils/trpc';
+import { Form, Loading, Select, Tooltip } from 'stratosphere-ui';
+import { trpc } from '../../../../utils/trpc';
 import { BAR_CHART_THEME } from './constants';
+import type { AirportsModeFormData } from './types';
 
-interface TopAircraftTypesFormData {
-  mode: 'flights' | 'distance' | 'duration';
-}
-
-export const TopAircraftTypesChart = (): JSX.Element => {
+export const TopAirportsChart = (): JSX.Element => {
   const { username } = useParams();
-  const methods = useForm<TopAircraftTypesFormData>({
+  const methods = useForm<AirportsModeFormData>({
     defaultValues: {
-      mode: 'flights',
+      mode: 'all',
     },
   });
-  const mode = useWatch<TopAircraftTypesFormData, 'mode'>({
+  const mode = useWatch<AirportsModeFormData, 'mode'>({
     name: 'mode',
     control: methods.control,
   });
-  const { data, isFetching } = trpc.statistics.getTopAircraftTypes.useQuery(
+  const { data, isFetching } = trpc.statistics.getTopAirports.useQuery(
     {
       username,
       limit: 5,
@@ -34,7 +31,7 @@ export const TopAircraftTypesChart = (): JSX.Element => {
   return (
     <div className="flex h-[180px] min-w-[250px] max-w-[500px] flex-1 flex-col items-center gap-1 font-semibold">
       <div className="flex h-9 w-full items-center justify-between">
-        <div className="text-sm">Top Aircraft Types</div>
+        <div className="text-sm">Top Airports</div>
         <Form methods={methods}>
           <Select
             buttonProps={{ color: 'ghost', size: 'xs' }}
@@ -42,20 +39,20 @@ export const TopAircraftTypesChart = (): JSX.Element => {
             getItemText={({ text }) => text}
             options={[
               {
-                id: 'flights',
-                text: 'Flights',
+                id: 'all',
+                text: 'All',
               },
               {
-                id: 'distance',
-                text: 'Distance (nm)',
+                id: 'departure',
+                text: 'Departure',
               },
               {
-                id: 'duration',
-                text: 'Duration (min)',
+                id: 'arrival',
+                text: 'Arrival',
               },
             ]}
             menuSize="sm"
-            menuClassName="w-[175px] right-0"
+            menuClassName="w-[185px] right-0"
             name="mode"
           />
         </Form>
@@ -77,8 +74,8 @@ export const TopAircraftTypesChart = (): JSX.Element => {
               theme={BAR_CHART_THEME}
               layout="horizontal"
               data={data}
-              keys={[mode]}
-              indexBy="aircraftType"
+              keys={['flights']}
+              indexBy="airport"
               enableGridY={false}
               axisBottom={{
                 tickSize: 0,
@@ -90,7 +87,16 @@ export const TopAircraftTypesChart = (): JSX.Element => {
               margin={{
                 left: 55,
               }}
-              colors={['var(--fallback-er,oklch(var(--er)/0.75))']}
+              colors={['var(--fallback-su,oklch(var(--su)/0.75))']}
+              tooltip={data => (
+                <Tooltip
+                  className="translate-y-[-20px]"
+                  open
+                  text={`${data.data.airport}: ${data.data.flights} ${
+                    data.data.flights > 1 ? 'flights' : 'flight'
+                  }`}
+                />
+              )}
             />
           </div>
         ) : null}
