+++ resources/pages/Profile/components/Statistics/index.ts
@@ -1,0 +1,1 @@
+export * from './Statistics';
