+++ resources/pages/Profile/components/Statistics/types.ts
@@ -1,0 +1,11 @@
+export type StatsTotalsMode = 'flights' | 'distance' | 'duration';
+
+export type StatsAirportMode = 'all' | 'departure' | 'arrival';
+
+export interface TotalsModeFormData {
+  mode: StatsTotalsMode;
+}
+
+export interface AirportsModeFormData {
+  mode: StatsAirportMode;
+}
