+++ resources/pages/Profile/components/index.ts
@@ -1,0 +1,7 @@
+export * from './AddFlight';
+export * from './CurrentFlight';
+export * from './Map';
+export * from './ProfileCard';
+export * from './RecentFlights';
+export * from './RouteTimeline';
+export * from './Statistics';
