+++ resources/pages/Profile/types.ts
@@ -1,0 +1,1 @@
+export type FlightDelayStatus = 'severe' | 'moderate' | 'none';
